refactor (core): refactor several components and Style
-Add refactor several style.
-Add folder utils.
