+++ app-vuelos/src/components/FloatingActionButton.js
@@ -1,36 +1,16 @@
-import { StyleSheet, Text, TouchableOpacity } from "react-native";
 import React from "react";
-import { Colors } from "../theme/Colors";
+import { Text, TouchableOpacity } from "react-native";
+import { stylesFloatinActionButton } from "./style/StyleFloatinActionButton";
 
 const FloatingActionButton = ({ navigation }) => {
   return (
     <TouchableOpacity
-      style={styles.containerFAB}
+      style={stylesFloatinActionButton.containerFAB}
       onPress={() => navigation.navigate("Booking")}
     >
-      <Text style={styles.title}>+</Text>
+      <Text style={stylesFloatinActionButton.title}>+</Text>
     </TouchableOpacity>
   );
 };
 
 export default FloatingActionButton;
-
-const styles = StyleSheet.create({
-  containerFAB: {
-    justifyContent: "center",
-    alignSelf: "center",
-    alignItems: "center",
-    position: "absolute",
-    bottom: 30,
-    backgroundColor: Colors.blue,
-    borderRadius: 70,
-    width: 80,
-    height: 80
-  },
-  title: {
-    fontSize: 45,
-    color: Colors.white,
-    fontWeight: "bold",
-    textAlign: "center"
-  }
-});
