+++ app-vuelos/src/components/Title.js
@@ -1,6 +1,5 @@
-import { StyleSheet, Text } from "react-native";
+import { Text } from "react-native";
 import React from "react";
-import { Colors } from "../theme/Colors";
 import { stylesTitle } from "./style/StyleTitle";
 
 const Title = ({ titleContent }) => {
@@ -8,5 +7,3 @@ const Title = ({ titleContent }) => {
 };
 
 export default Title;
-
-
