+++ app-vuelos/src/navigations/RootNativagtions.js
@@ -1,6 +1,4 @@
 import React from "react";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-
 //import Views
 import { NavigationContainer } from "@react-navigation/native";
 import MyStack from "./MyStack";
