+++ app-vuelos/src/views/style/StyleBooking.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+import { Colors } from "../../theme/Colors";
+
+export const stylesBoking = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center"
+  },
+  txt: {
+    fontSize: 30,
+    fontWeight: "bold",
+    color: Colors.red
+  }
+});
